+++ sdk/resourcemanager/azure-resourcemanager-network/CHANGELOG.md
@@ -8,6 +8,8 @@
 
 ### Bugs Fixed
 
+- Fixed bug that "Microsoft.Network/bastionHosts" resource causes exception on `listNetworkInterfaceIPConfigurations` method in `Subnet` class.
+
 ### Other Changes
 
 ## 2.44.0 (2024-10-25)